requirements.txt
altair
annotated-types
anthropic
anyio
asttokens
attrs
audio-recorder-streamlit
blinker
cachetools
certifi
charset-normalizer
click
colorama
comm
debugpy
decorator
distro
entrypoints
executing
filelock
fsspec
gitdb
GitPython
google-ai-generativelanguage
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-generativeai
googleapis-common-protos
google-cloud-speech
google-cloud-texttospeech
grpcio
grpcio-status
h11
httpcore
httplib2
httpx
huggingface-hub
idna
importlib-metadata
IProgress
ipykernel
ipython
jedi
Jinja2
jiter
jsonschema
jsonschema-specifications
jupyter_client
jupyter_core
markdown-it-py
MarkupSafe
matplotlib-inline
mdurl
nest-asyncio
numpy
openai
packaging
pandas
parso
pillow
platformdirs
prompt-toolkit
proto-plus
protobuf
psutil
pure-eval
pyarrow
pyasn1
pyasn1_modules
pydantic
pydantic_core
pydeck
Pygments
pyparsing
python-dateutil
python-dotenv
pytz
PyYAML
pyzmq
referencing
requests
rich
rpds-py
rsa
six
smmap
sniffio
stack-data
streamlit
tenacity
tokenizers
toml
toolz
tornado
tqdm
traitlets
typing_extensions
tzdata
tzlocal
uritemplate
urllib3
validators
watchdog
wcwidth
zipp
